Moved the event bus to the view model
