+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/details/OrderDetailFragment.kt
@@ -46,7 +46,6 @@ import com.woocommerce.android.ui.orders.AddOrderShipmentTrackingFragment
 import com.woocommerce.android.ui.orders.OrderNavigationTarget
 import com.woocommerce.android.ui.orders.OrderNavigator
 import com.woocommerce.android.ui.orders.OrderProductActionListener
-import com.woocommerce.android.ui.orders.details.OrderDetailRepository.OnProductImageChanged
 import com.woocommerce.android.ui.orders.details.adapter.OrderDetailShippingLabelsAdapter.OnShippingLabelClickListener
 import com.woocommerce.android.ui.orders.notes.AddOrderNoteFragment
 import com.woocommerce.android.ui.orders.shippinglabels.ShippingLabelRefundFragment
@@ -57,9 +56,6 @@ import com.woocommerce.android.viewmodel.ViewModelFactory
 import com.woocommerce.android.widgets.SkeletonView
 import dagger.android.support.AndroidSupportInjection
 import kotlinx.android.synthetic.main.fragment_order_detail.*
-import org.greenrobot.eventbus.EventBus
-import org.greenrobot.eventbus.Subscribe
-import org.greenrobot.eventbus.ThreadMode.MAIN
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.order.CoreOrderStatus
 import javax.inject.Inject
 
@@ -95,12 +91,10 @@ class OrderDetailFragment : BaseFragment(), NavigationResult, OrderProductAction
     override fun onResume() {
         super.onResume()
         AnalyticsTracker.trackViewShown(this)
-        EventBus.getDefault().register(this)
     }
 
     override fun onStop() {
         undoSnackbar?.dismiss()
-        EventBus.getDefault().unregister(this)
         super.onStop()
     }
 
@@ -146,6 +140,7 @@ class OrderDetailFragment : BaseFragment(), NavigationResult, OrderProductAction
             new.isRefreshing?.takeIfNotEqualTo(old?.isRefreshing) {
                 orderRefreshLayout.isRefreshing = it
             }
+            new.refreshedProductId?.takeIfNotEqualTo(old?.refreshedProductId) { refreshProduct(it) }
         }
 
         viewModel.orderNotes.observe(viewLifecycleOwner, Observer {
@@ -224,6 +219,10 @@ class OrderDetailFragment : BaseFragment(), NavigationResult, OrderProductAction
 
     private fun showOrderNotesSkeleton(show: Boolean) {
         orderDetail_noteList.showSkeleton(show)
+    }
+
+    private fun refreshProduct(remoteProductId: Long) {
+        orderDetail_productList.notifyProductChanged(remoteProductId)
     }
 
     private fun showOrderNotes(orderNotes: List<OrderNote>) {
@@ -369,11 +368,5 @@ class OrderDetailFragment : BaseFragment(), NavigationResult, OrderProductAction
             it.addCallback(dismissCallback)
             it.show()
         }
-    }
-
-    @SuppressWarnings("unused")
-    @Subscribe(threadMode = MAIN)
-    fun onProductImageChanged(event: OnProductImageChanged) {
-        orderDetail_productList.notifyProductChanged(event.remoteProductId)
     }
 }
